Stepper and Beam Control
Stepper:
Back to normal stepper library with low revolution count to enable much faster speeds. This along with direct 9v plug in works great.

Laser:
Took laser apart and connected to to the Arduino via a PNP transistor switch, allow it to be rapidly turned on and off in sync with the stepper motor. Thus, removes the ghosting the beams previously had with the light constantly turned on.